refactor(native): Jinja - separate injection of Python filters
chore: fix
